Refactor survey tests to fix SwiftLint violations
Split comprehensive survey tests into separate files to comply with
file and type length limits. Moved real-world survey scenarios to
dedicated test suite and extracted test helpers to separate file.
